Moved head sync functionality to CharacterOutfitSync
