+++ Assets/Scripts/CharacterController/CharacterController.cs
@@ -5,9 +5,7 @@ using UnityEngine;
 
 public class CharacterController : NetworkBehaviour
 {
-    [SerializeField] private HumanoidRigidRig rigParts;
-
-    [SerializeField] private HeadPrefabs headPrefabs;
+    public HumanoidRigidRig rigParts;
 
     private Rigidbody rb;
 
@@ -19,45 +17,6 @@ public class CharacterController : NetworkBehaviour
     private ProceduralAnimation<HumanoidRigidRig> walkAnim;
     private ProceduralAnimation<HumanoidRigidRig> idleAnim;
     private ProceduralAnimation<HumanoidRigidRig> attackAnim;
-
-    // ---Head sync functionality---
-    [SyncVar(hook = nameof(SetGender))]
-    int genderIndex = 0;
-
-    [SyncVar(hook = nameof(SetRace))]
-    int raceIndex = 0;
-
-    [Command]
-    public void CmdSetGender(int _genderIndex)
-    {
-        this.genderIndex = _genderIndex;
-    }
-
-    [Command]
-    public void CmdSetRace(int _raceIndex)
-    {
-        this.raceIndex = _raceIndex;
-    }
-
-    [Command]
-    public void CmdSetGenderAndRace(int _genderIndex, int _raceIndex)
-    {
-        this.genderIndex = _genderIndex;
-        this.raceIndex = _raceIndex;
-    }
-
-    // Clientside SyncVar hook
-    void SetGender(int oldGender, int newGender)
-    {
-        ChangeHeadTo(newGender, raceIndex);
-    }
-
-    // Clientside SyncVar hook
-    void SetRace(int oldRace, int newRace)
-    {
-        ChangeHeadTo(genderIndex, newRace);
-    }
-    // End of ---Head sync functionality---
 
     void Start()
     {
@@ -72,13 +31,8 @@ public class CharacterController : NetworkBehaviour
             cif = new LocalKeyboardCIF(camController);
             camController.SetCameraTarget(transform);
             HumanoidRigInitialPose.SetupInstance(rigParts);
-
-            raceIndex = PlayerPrefs.GetInt("RaceSelected", 0);
-            genderIndex = PlayerPrefs.GetInt("GenderSelected", 0);
-
-            CmdSetGenderAndRace(genderIndex, raceIndex);
-
-            ChangeHeadTo(genderIndex, raceIndex);
+            
+            GetComponent<CharacterOutfitSync>().LocalInit();
         }
         else
         {
@@ -95,22 +49,6 @@ public class CharacterController : NetworkBehaviour
         //animationController.SwitchTo(attackAnim);
 
         movementController = new MovementController(rb, cif);
-    }
-
-    int genderCache = 0;
-    int raceCache = 0;
-    public void ChangeHeadTo(int gender, int race)
-    {
-        if (gender == genderCache && race == raceCache) { return; }
-
-        genderCache = gender;
-        raceCache = race;
-
-        GameObject newHeadPrefab = headPrefabs.GetHead((Gender)gender, (CharacterRace)race);
-
-        GameObject newHead = Instantiate(newHeadPrefab, transform);
-        Destroy(rigParts.head.gameObject);
-        rigParts.head = newHead.transform;
     }
 
     private void Update()
