+++ Assets/Scripts/CharacterController/CharacterOutfitSync.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b79ec97cb7312084b9bc344a32c8ed9b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
